Disabled the F5 key to refresh
Pressing 'F5' no longer refreshes the file panels. This is for a simple
reason: holding down F5 would cause the app to crash because Textual did
not like how many times the function was being called. A simple solution
would be to introduce a rate limit to the button, but refreshing as soon
as an action is taken is MUCH simpler and a more automatic process in
the first place. The refresh action may be re-added in a future release,
with a 500ms rate limit.